requirements.txt
torch
faster-whisper
fastapi
uvicorn
ffmpeg-python
